pass params by console or file ini  (#120)
* first test DANGEROUS ! mount root auto

* universal read input

* format disk by param

* auto load root partition

* simplify structure

* getvar() function

+++ lib/util-disk.sh
@@ -235,17 +235,22 @@ select_filesystem() {
     fs_opts=""
     CHK_NUM=0
 
-    DIALOG " $_FSTitle " --menu "\n$_FSBody\n " 0 0 12 \
-      "$_FSSkip" "-" \
-        "btrfs" "mkfs.btrfs -f" \
-        "ext3" "mkfs.ext3 -q" \
-        "ext4" "mkfs.ext4 -q" \
-        "jfs" "mkfs.jfs -q" \
-        "nilfs2" "mkfs.nilfs2 -fq" \
-        "ntfs" "mkfs.ntfs -q" \
-        "reiserfs" "mkfs.reiserfs -q" \
-        "vfat" "mkfs.vfat -F32" \
-        "xfs" "mkfs.xfs -f" 2>${ANSWER} || return 1
+    local option==$(getvar "mount.${PARTITION}")
+    if [[ -z "$option" ]]; then
+        DIALOG " $_FSTitle " --menu "\n${PARTITION}\n$_FSBody\n " 0 0 12 \
+        "$_FSSkip" "-" \
+            "btrfs" "mkfs.btrfs -f" \
+            "ext3" "mkfs.ext3 -q" \
+            "ext4" "mkfs.ext4 -q" \
+            "jfs" "mkfs.jfs -q" \
+            "nilfs2" "mkfs.nilfs2 -fq" \
+            "ntfs" "mkfs.ntfs -q" \
+            "reiserfs" "mkfs.reiserfs -q" \
+            "vfat" "mkfs.vfat -F32" \
+            "xfs" "mkfs.xfs -f" 2>${ANSWER} || return 1
+    else
+        echo "$option">${ANSWER}
+    fi
         
     case $(cat ${ANSWER}) in
         "$_FSSkip") FILESYSTEM="$_FSSkip"
@@ -840,8 +845,11 @@ mount_partitions() {
     done
 
     # Identify and mount root
-    DIALOG " $_PrepMntPart " --menu "\n$_SelRootBody\n " 0 0 12 ${PARTITIONS} 2>${ANSWER} || return 0
-    PARTITION=$(cat ${ANSWER})
+    PARTITION=$(getvar "mount.root")
+    if [[ -z "$PARTITION" ]]; then
+        DIALOG " $_PrepMntPart " --menu "\n$_SelRootBody\n " 0 0 12 ${PARTITIONS} 2>${ANSWER} || return 0
+        PARTITION=$(cat ${ANSWER})
+    fi
     ROOT_PART=${PARTITION}
 
     # Format with FS (or skip) -> # Make the directory and mount. Also identify LUKS and/or LVM
